fix: add validation to ensure total duration does not exceed 24 hours in claim form
